preserve the fractional track and speed the wire carries
SBS-1 emits decimal track and ground speed ("270.5", "145.6") but both
parsed via strtoul, which stops at the dot: the report sent 27000
centidegrees when the wire had 27050, paying for centidegree/cm-s
resolution in the FSS message and discarding it at parse time.

Parse both as double through the same strict from_chars helper as
coordinates, store them as doubles on the record, and round (not
truncate) in the report conversions. The knots factor is now written
exactly as 1852.0/36.0 rather than the hand-truncated 51.444, which was
already off by one cm/s at 450 kt.

Track is range-checked to [0, 360]: a corrupt value like 70000 used to
be clamped to UINT16_MAX and reported as a plausible-looking heading;
it now leaves the heading unset, consistent with the coordinate policy.

+++ src/dump1090.hpp
@@ -23,9 +23,12 @@ private:
     bool callsign_set{false};
     uint32_t altitude{0};
     bool altitude_set{false};
-    uint32_t speed{0};
+    /* Speed (knots) and heading (degrees) are stored as the wire carries
+     * them — with a decimal fraction — so the report-time conversion to cm/s
+     * and centidegrees keeps the resolution those units pay for. */
+    double speed{0.0};
     bool speed_set{false};
-    uint16_t heading{0};
+    double heading{0.0};
     bool heading_set{false};
     Point pos{};
     int16_t vert_vel{0};
@@ -52,19 +55,19 @@ public:
     };
     [[nodiscard]] auto getAltitude() const -> uint32_t { return this->altitude; };
     [[nodiscard]] auto validAltitude() const -> bool { return this->altitude_set; };
-    void setSpeed(uint32_t t_speed)
+    void setSpeed(double t_speed)
     {
         this->speed = t_speed;
         this->speed_set = true;
     };
-    [[nodiscard]] auto getSpeed() const -> uint32_t { return this->speed; };
+    [[nodiscard]] auto getSpeed() const -> double { return this->speed; };
     [[nodiscard]] auto validSpeed() const -> bool { return this->speed_set; };
-    void setHeading(uint16_t t_heading)
+    void setHeading(double t_heading)
     {
         this->heading = t_heading;
         this->heading_set = true;
     };
-    [[nodiscard]] auto getHeading() const -> uint16_t { return this->heading; };
+    [[nodiscard]] auto getHeading() const -> double { return this->heading; };
     [[nodiscard]] auto validHeading() const -> bool { return this->heading_set; };
     void setVertVel(int16_t t_vert_vel)
     {
